LayoutTests/canvas/philip/tests/2d.path.stroke.empty-expected.txt

< [index] >
2d.path.stroke.empty
Empty subpaths are not stroked
References: 2d.path.stroke.empty
Actual output:
Expected output:

Passed
